refactor(tauri): move the startup helpers out of lib.rs (Unit 11f)
Verbatim move of 856 lines. The startup-file rejection buffer and its test
module go to a new `startup_rejection.rs`; the autostart launch-mode helpers
and `autostart_tests` join the registration commands they already belong to in
the existing `autostart.rs`. `lib.rs` 5101 -> 4271 (856 out, 26 back in as the
crate-root re-exports and their rationale), and `run` keeps the whole plugin
chain.

Two clusters and two destinations rather than one `startup.rs`, because
"startup" is a phase and every other module this epic produced is named for a
feature -- and because `autostart.rs` already existed, so a new module would
have split one feature across three files.

`lib.rs` re-exports the moved names at the crate root, the idiom it already
uses for `open_candidate`, instead of qualifying ~20 call sites. That keeps
every call site byte-identical, keeps the `{AUTOSTART_FLAG}` format-string
captures working, and -- the reason it is not merely tidier -- keeps a sibling
module's `use crate::{...}` resolving across the split. #1651's `sidecar.rs`
imports three of these names; qualifying instead would have compiled here and
broken a file git merges clean, with no conflict marker to warn anyone.

Verified mechanically against `git show HEAD:lib.rs` in BOTH directions: 13
lines differ and all 13 are the `pub(crate)` the move requires, and zero
uniquely-identifying moved lines survive in `lib.rs`. The reverse check earned
its place -- a mutant that leaves a duplicate buffer behind compiles, passes
all 287 Rust tests and every TS guard, and only that check sees it. `#[test]`
count is conserved exactly, 77 to 77.

No `pub(crate)` was added to `lib.rs`: a private crate-root item is visible to
every descendant module, so the moved tests reach `LAUNCHER_DEFERRED` and
`post_paths_and_surface` unchanged. `has_argv_flag` and `LAUNCHER_DEFERRED`
stay at the crate root -- both have consumers outside this cluster.

Guard work, each measured:
- The two hardcoded `lib.rs` reads become `rustSourceDefining` locators, and
  now read `code` rather than `text`. Moving the locator alone would have left
  the defeat where 11c and 11d each found it: a `#[cfg(test)]` fake of the
  gave-up arm satisfies a `text` scan while the real arm is gutted -- green
  before the switch, red after, with the switch reverted to prove it.
- The decline-arm spec is sliced from `show_server_error_dialog` instead of
  taking the file's first `if !retry`.
- New: Rust's `EVENT_STARTUP_FILE_REJECTED` must equal the client's
  `NUDGE_EVENT`. The parity check keys on `CODE_*`, so the event name sat
  outside every gate in the repo -- renaming it is green everywhere while
  every warm-path toast silently stops arriving.
- The walk control is derived ("the file declaring the wire codes is in the
  walk") rather than naming the new module, so a harmless rename stays green.

`cargo fmt` deliberately not run; the repo is not rustfmt-clean.
