+++ src-tauri/src/lib.rs
@@ -5,6 +5,7 @@ mod native_theme;
 mod pending_update;
 pub mod keychain;
 mod sentry_reporting;
+mod startup_rejection;
 mod token_store;
 mod uninstall_scrub;
 
@@ -76,12 +77,42 @@ pub use open_candidate::{
 // `single-instance` callback and cold start). The other two are reached only
 // from the macOS-gated `handle_opened_urls`, so their import carries the same
 // cfg as the caller rather than an `allow(unused_imports)` — `test` is in the
-// set because `startup_rejection_tests` exercises them on every CI leg.
+// set because `startup_rejection_tests` exercises them on every CI leg. That
+// module moved to `startup_rejection.rs` with Unit 11f and reaches these
+// through `crate::`, so the `test` arm is still load-bearing and still earned
+// by that one module: `classify_opened_url_tests`, the test module that stayed,
+// uses `classify_opened_url` and `OpenedUrlRejection` but never
+// `opened_url_reason_code`.
 pub(crate) use open_candidate::rejection_reason_code;
 #[cfg(any(target_os = "macos", test))]
 pub(crate) use open_candidate::{classify_opened_url, opened_url_reason_code, OpenedUrlRejection};
 #[cfg(test)]
 pub(crate) use open_candidate::validate_open_candidate;
+
+// Re-exported rather than qualified at ~20 call sites, the same idiom as
+// `open_candidate` above and for a stronger reason: `use crate::{…}` in a
+// sibling module is how the rest of the crate already reaches these names, so
+// re-exporting keeps every call site — here and in files this unit does not
+// touch — byte-identical across the split. Qualifying instead would compile
+// here and break a sibling that git merged clean, with no conflict to warn
+// anyone. `get_startup_rejection` is deliberately absent: `#[tauri::command]`
+// generates a sibling `__cmd__…` macro that a plain `use` of the function does
+// not import, so `generate_handler!` module-qualifies it as 11b–11d did.
+pub(crate) use startup_rejection::{
+    clear_startup_rejection, surface_startup_rejection, RejectionBatch, CODE_MULTIPLE_DEFERRED,
+    CODE_OPEN_DEFERRED, CODE_OPEN_FAILED,
+};
+// The launch-mode half of start-at-login moved into `autostart.rs` alongside
+// the registration commands. `autostart_seen_and_mark` carries its item's own
+// `#[cfg(target_os = "linux")]`: an ungated re-export of a gated item is an
+// unconditional E0432 on the other two platforms, which is the shape that broke
+// Unit 11c and that only CI can see.
+pub(crate) use autostart::{
+    is_autostart_launch, resolve_autostart_launch, should_start_hidden, AUTOSTART_DISABLE_ENV,
+    AUTOSTART_FLAG,
+};
+#[cfg(target_os = "linux")]
+pub(crate) use autostart::autostart_seen_and_mark;
 
 // Bare `PathBuf` survives only in the `#[cfg(test)]` modules now. Every
 // unconditionally-compiled use went to `open_candidate.rs` with the
@@ -243,252 +274,6 @@ static SIDECAR_GAVE_UP: AtomicBool = AtomicBool::new(false);
 /// the rest of the session.
 static LAUNCHER_DEFERRED: AtomicBool = AtomicBool::new(false);
 
-/// Buffered file-open rejection reason CODE (stable, path-free) — the single
-/// delivery surface for EVERY entry point, cold start and warm start alike.
-/// See issues #630 and #1344.
-///
-/// ## Why buffer instead of just emitting an event
-///
-/// A cold-start "Open With" rejection (`extract_file_arg` returns `Err`) is
-/// classified in `setup()` — which runs before `App.svelte` exists at all, let
-/// alone its listener. Emitting a Tauri event there drops silently on the exact
-/// failure mode it is meant to surface. So the reason is buffered here and
-/// drained via `get_startup_rejection()` when the client comes up.
-///
-/// ## Why the buffer is unconditional, and not gated on WebView readiness
-///
-/// `startup-file-rejected` is a payload-free NUDGE, answered by calling
-/// `get_startup_rejection()` — the same take-once accessor the init-time drain
-/// uses. So Rust never has to know whether the listener is wired, which it
-/// cannot know.
-///
-/// An earlier revision of this work (see PR #1414's history; the design does not
-/// appear in issue #1344 itself) inferred readiness from a `REJECTION_POLLED`
-/// latch — "the WebView has polled, so its listener must be wired" — and skipped
-/// the buffer once set. The inference does not hold, for a reason that needs no
-/// race to demonstrate: the latch is process-global and the listener is not. **A
-/// WebView reload leaves the latch set and the listener gone**, so every
-/// rejection after the first reload took the emit-only path and was dropped
-/// permanently. (`App.svelte` also wires the listener and drains in two
-/// independent promise chains whose completion order is not guaranteed, so a
-/// completed drain never proved a wired listener either — but that one is a
-/// sub-millisecond window and is not the argument this rests on.)
-///
-/// Unconditional buffering also means a nudge with no listener costs one dropped
-/// event rather than the toast. Same queue-then-drain shape as `PendingOpens`,
-/// the other place in this file where a producer must not care whether the
-/// consumer is ready.
-///
-/// ## Why a `Mutex<Option<_>>` and not a `OnceLock`
-///
-/// `OnceLock` cannot be cleared, but the buffer MUST be cleared on
-/// `restart_sidecar` so a stale rejection from a previous launch isn't replayed
-/// against the new sidecar. A `Mutex<Option<String>>` gives set / take / clear.
-///
-/// ## No path leakage
-///
-/// The buffer holds a stable reason CODE only (e.g. `"unsupported-extension"`),
-/// never the rejected path — the resolved path is already logged at `warn` for
-/// diagnostics, and the human-readable toast message is composed client-side
-/// (mirrors the path-free `sidecar-restart-failed` toast contract). The writers
-/// take `&'static str` precisely so this cannot be relaxed by accident; see
-/// [`buffer_startup_rejection_code`].
-static STARTUP_REJECTION: Mutex<Option<String>> = Mutex::new(None);
-
-/// Run `f` against the rejection buffer, recovering from (and reporting) a
-/// poisoned mutex. One helper rather than three hand-rolled `match` arms: the
-/// take path used to be the only one that recovered *silently*, and it is the
-/// one whose recovery destroys state, so the divergence was exactly backwards.
-fn with_rejection<R>(what: &str, f: impl FnOnce(&mut Option<String>) -> R) -> R {
-    match STARTUP_REJECTION.lock() {
-        Ok(mut guard) => f(&mut guard),
-        Err(poisoned) => {
-            log::error!("STARTUP_REJECTION mutex poisoned during {what} — recovering");
-            f(&mut poisoned.into_inner())
-        }
-    }
-}
-
-/// Tauri event name for a startup-file rejection surfaced to the WebView.
-/// Deliberately payload-FREE: it is a nudge to call `get_startup_rejection`,
-/// not the delivery itself (see [`STARTUP_REJECTION`]). Keeping the code out of
-/// the payload is what makes the buffer the single source of truth, so a nudge
-/// with no listener costs nothing.
-const EVENT_STARTUP_FILE_REJECTED: &str = "startup-file-rejected";
-
-/// Record an already-mapped, path-free reason CODE in the buffer. The single
-/// writer — every entry point funnels through [`surface_startup_rejection`],
-/// which funnels through here.
-///
-/// `&'static str`, not `&str`, and that is a safety property rather than a
-/// style choice. `Display for OpenedUrlRejection` deliberately EMBEDS the
-/// resolved path (it is the log format), so a sink taking `&str` would accept
-/// `&format!("{reason}")` and put a filesystem path on the wire to the DOM.
-/// A formatted string is not `'static`, so that mistake is now a compile error
-/// rather than a thing tests have to notice. Both producers
-/// ([`rejection_reason_code`], [`opened_url_reason_code`]) already return
-/// `&'static str` from a closed set of literals.
-///
-/// Last-write-wins. A multi-file Opened batch rejects at most once through
-/// [`surface_startup_rejection`] — the batch collapses to a single code before
-/// it reaches here — so the surviving value is one deliberate summary, not an
-/// arbitrary survivor of N racing writes.
-fn buffer_startup_rejection_code(code: &'static str) {
-    with_rejection("buffer", |slot| *slot = Some(code.to_string()));
-}
-
-/// Deliver a startup-file rejection to the WebView. THE delivery path — every
-/// OS entry point that rejects a candidate calls this and nothing else: argv
-/// cold start, second-instance warm start, and the macOS Apple Event, which
-/// alone serves cold and warm start through one arm.
-///
-/// Buffer first, then nudge, unconditionally. Do not re-add a "has the WebView
-/// polled yet" discriminator to skip the buffer on warm start — [`STARTUP_REJECTION`]
-/// records why the readiness inference is false (it also raced, being read
-/// outside the buffer's mutex).
-fn surface_startup_rejection(app: &tauri::AppHandle, code: &'static str) {
-    surface_startup_rejection_with(code, || {
-        if let Err(e) = app.emit(EVENT_STARTUP_FILE_REJECTED, ()) {
-            // Non-fatal by construction: the code is already buffered, so the
-            // worst case is that the toast waits for the next client init.
-            log::warn!("Failed to emit {EVENT_STARTUP_FILE_REJECTED}: {e}");
-        }
-    });
-}
-
-/// The testable half of [`surface_startup_rejection`], split out because
-/// `tauri::AppHandle` cannot be constructed in a unit test (no `tauri` `test`
-/// feature is in dev-dependencies) — and without this seam the ordering below
-/// was pinned by nothing at all: deleting the buffer call reverted the whole
-/// fix, and swapping the two lines re-opened the drop, both with a green suite.
-///
-/// The order is the contract. Buffer BEFORE nudging, or a live listener drains
-/// an empty slot and the toast is lost — the nudge is deliberately payload-free,
-/// so it carries nothing the client could fall back on.
-fn surface_startup_rejection_with(code: &'static str, nudge: impl FnOnce()) {
-    buffer_startup_rejection_code(code);
-    nudge();
-}
-
-/// Wire code for "more than one file in this batch was refused". Distinct from
-/// the per-reason codes because a mixed batch has no single true reason, and
-/// picking one of them would state something false rather than something vague.
-///
-/// Ungated since #1416: `post_paths_and_surface` gives it callers on every
-/// platform.
-const CODE_MULTIPLE_REJECTED: &str = "multiple-rejected";
-
-/// Wire code for "the candidate passed shell validation, and then failed to open
-/// anyway" — a non-2xx from `/api/open`, a transport failure, or an open that
-/// arrived after the app stopped trying to start the server (#1416).
-///
-/// Deliberately ONE code for all three, not three: the user's remedy is identical
-/// (try again / reopen the file), the distinction is a diagnostic `tandem.log`
-/// already records with the real error text, and every extra code is another
-/// string that can desync from a stale WebView. Distinct from the *validation*
-/// codes because nothing about the file was wrong — `messageForStartupRejection`
-/// maps it to the deliberately vague "That file couldn't be opened in Tandem."
-const CODE_OPEN_FAILED: &str = "open-failed";
-
-/// Wire codes for "queued, not yet delivered, and the queue is RETAINED".
-///
-/// Split from [`CODE_OPEN_FAILED`] because that code's message is flatly past
-/// tense — "That file couldn't be opened in Tandem." — and for a retained queue
-/// that is a false statement: `promote_healthy_and_drain` will open the file if
-/// the user ever restarts the server, which the failure dialog explicitly tells
-/// them how to do. Asserting a finality the design does not have is the same
-/// defect as the "Abandoning N queued file open(s)" line deleted in this change,
-/// and it would be read by the one user in a position to act on it.
-///
-/// Two codes rather than one because the singular message cannot describe a
-/// five-file drop, mirroring the [`CODE_MULTIPLE_REJECTED`] split.
-const CODE_OPEN_DEFERRED: &str = "open-deferred";
-const CODE_MULTIPLE_DEFERRED: &str = "multiple-deferred";
-
-/// Collapses one OS batch — a Finder multi-select "Open With", a multi-file Dock
-/// drop — into the single code the user should see.
-///
-/// This exists because the buffer is one slot and the batch is N rejections. The
-/// obvious shape (call [`surface_startup_rejection`] per rejected URL) is what
-/// the first version did, and it is nondeterministic in a way that shows: the
-/// loop is synchronous while the client's drain is an async IPC round-trip, so
-/// whether the user got one toast or two-with-a-count-badge, and which reason it
-/// named, depended on where the drain happened to land. Same user action, two
-/// different outcomes.
-///
-/// Accumulating first and surfacing once removes the race rather than narrowing
-/// it. Deliberately ungated and free of Tauri types so it is unit-testable on
-/// every platform — its principal caller, `handle_opened_urls`, is macOS-only and
-/// cannot be tested from Windows or Linux CI, which is precisely why the logic
-/// must not live inside it.
-///
-/// Since #1416 the batch also carries POST failures (`CODE_OPEN_FAILED`), so ONE
-/// accumulator spans validation *and* delivery for a single OS batch. That is
-/// what keeps "one batch, one surface call" true **for everything that resolves
-/// synchronously with the batch**: a Finder multi-select of a `.pdf` (refused)
-/// plus a 60 MB `.md` (refused by the server) would otherwise write twice into
-/// the one-slot buffer and the user would see a count badge whose value depended
-/// on where the client's async drain landed.
-///
-/// **The claim stops at the async boundary, deliberately.** A path that returns
-/// `OpenRoute::Queued` has no outcome yet, and may not have one for minutes —
-/// its verdict arrives from `report_pending_opens_with` when the user answers
-/// the server-failure dialog. Carrying the accumulator across that gap was
-/// considered and rejected: it would have to stay open for a user-timescale
-/// wait, trading a merge bug for a staleness bug. Instead the client keys its
-/// toast dedup on the CODE (`startup-file-rejected:${code}`), so a validation
-/// reason and a later delivery verdict can never merge into one count badge.
-#[derive(Default)]
-struct RejectionBatch {
-    first: Option<&'static str>,
-    count: usize,
-}
-
-impl RejectionBatch {
-    fn record(&mut self, code: &'static str) {
-        self.count += 1;
-        if self.first.is_none() {
-            self.first = Some(code);
-        }
-    }
-
-    /// The one code to surface, or `None` when nothing was rejected. A single
-    /// rejection keeps its specific reason; two or more report multiplicity,
-    /// because "that file type can't be opened" is actively misleading when the
-    /// user just dropped five files and four of them opened.
-    fn resolve(&self) -> Option<&'static str> {
-        match self.count {
-            0 => None,
-            1 => self.first,
-            _ => Some(CODE_MULTIPLE_REJECTED),
-        }
-    }
-}
-
-/// Clear any buffered rejection. Called from `restart_sidecar` so a stale
-/// rejection from the previous launch can't be replayed against the new sidecar.
-///
-/// Note the scope is wider than that rationale: since every entry point buffers,
-/// this also discards a LIVE rejection caught in the window between
-/// [`surface_startup_rejection`] and the client's drain. That needs the user to
-/// hit Relaunch in the same breath as a rejected open, and the alternative
-/// (replaying it against the new sidecar) is worse.
-fn clear_startup_rejection() {
-    with_rejection("clear", |slot| *slot = None);
-}
-
-/// Client-drained accessor for the buffered rejection code. Returns `Some(code)`
-/// exactly once per buffered rejection: the value is TAKEN, so a re-init (e.g.
-/// an in-WebView reload) doesn't replay a toast the user already saw. Called
-/// from BOTH `App.svelte` sites — the init-time drain and the
-/// `startup-file-rejected` listener — because the event is a payload-free nudge
-/// and this is the only way to read what it is about. Take-once is what makes
-/// that safe: whichever arrives first wins, the other gets `None`.
-#[tauri::command]
-fn get_startup_rejection() -> Option<String> {
-    with_rejection("take", |slot| slot.take())
-}
-
 /// Strip the Windows extended-length path prefix (`\\?\`) that Tauri's
 /// `resource_dir()` / `app_data_dir()` return. Node.js can't resolve these.
 fn strip_win_prefix(path: &std::path::Path) -> String {
@@ -505,23 +290,6 @@ const MENU_UPDATE: &str = "update";
 
 const MAIN_WINDOW_LABEL: &str = "main";
 
-/// argv flag the OS autostart registration carries, so a login-triggered launch
-/// is distinguishable from a user-initiated one. An argv flag rather than an env
-/// var because the registration itself (Run value / plist / .desktop Exec line)
-/// is the only thing the OS lets us control, and it survives being inspected in
-/// Task Manager → Startup where an env var would be invisible.
-///
-/// `extract_file_arg` already skips `-`-prefixed args, so this can never be
-/// mistaken for a file-association path — `autostart_flag_is_not_a_file_arg`
-/// pins that.
-pub(crate) const AUTOSTART_FLAG: &str = "--tandem-autostart";
-
-/// Debug escape hatch: forces a launch to behave as a normal one (always show
-/// the window, never defer the Claude launcher) even when the OS passed
-/// `--tandem-autostart`. Mirrors `TANDEM_DISABLE_LAUNCHER` in spirit — every
-/// other lifecycle behavior here has an env opt-out.
-const AUTOSTART_DISABLE_ENV: &str = "TANDEM_DISABLE_AUTOSTART";
-
 /// Exact-match argv flag predicate, skipping `argv[0]`.
 ///
 /// The skip is a security invariant, not a nicety: an executable literally
@@ -530,68 +298,6 @@ const AUTOSTART_DISABLE_ENV: &str = "TANDEM_DISABLE_AUTOSTART";
 /// can't copy the invariant a third time and get it subtly wrong.
 pub(crate) fn has_argv_flag(args: &[String], flag: &str) -> bool {
     args.iter().skip(1).any(|a| a == flag)
-}
-
-/// True when this process was started by the OS at login.
-pub(crate) fn is_autostart_launch(args: &[String]) -> bool {
-    has_argv_flag(args, AUTOSTART_FLAG)
-}
-
-/// Resolve the effective autostart state for this process: the flag, minus the
-/// env kill switch. Deliberately does not log — it is called before the log
-/// plugin is registered (see the `setup()` ordering comment), so the one
-/// interesting case (the override actually firing) is logged at the call site
-/// once logging is live.
-fn resolve_autostart_launch(args: &[String], disable_env: Option<&str>) -> bool {
-    if disable_env == Some("1") {
-        return false;
-    }
-    is_autostart_launch(args)
-}
-
-/// Whether a boot launch should stay hidden in the tray.
-///
-/// The tray guard is load-bearing, not cosmetic: on Linux without
-/// libappindicator `TrayIconBuilder::build()` fails and `CloseRequested` exits
-/// the process instead of hiding. A hidden window with no tray icon would be an
-/// unreachable zombie still holding :3478/:3479 with no way to quit it short of
-/// a task manager.
-fn should_start_hidden(autostart: bool, tray_available: bool) -> bool {
-    autostart && tray_available
-}
-
-/// Marker recording that at least one autostart launch has already happened.
-/// Lives beside the session data in the app data dir.
-#[cfg(target_os = "linux")]
-const AUTOSTART_SEEN_MARKER: &str = "autostart-seen";
-
-/// Linux-only backstop for the residual hole in `should_start_hidden`.
-///
-/// `TrayIconBuilder::build()` *succeeding* does not prove the icon is visible —
-/// on GNOME without a status-icon extension it constructs fine and renders
-/// nothing, which is exactly the unreachable-process case the tray guard exists
-/// to prevent. So the first-ever autostart launch always shows the window,
-/// giving the user one guaranteed chance to see Tandem running and turn the
-/// setting off. Subsequent launches trust the tray.
-///
-/// Returns whether a prior autostart launch was recorded, and writes the marker
-/// as a side effect. Any I/O failure is reported as "not seen", which fails
-/// toward showing the window — always recoverable, unlike failing toward hidden.
-#[cfg(target_os = "linux")]
-fn autostart_seen_and_mark(dir: &std::path::Path) -> bool {
-    let marker = dir.join(AUTOSTART_SEEN_MARKER);
-    if marker.exists() {
-        return true;
-    }
-    if let Err(e) = std::fs::create_dir_all(dir) {
-        log::warn!("Could not create app data dir for autostart marker: {e}");
-        return false;
-    }
-    if let Err(e) = std::fs::write(&marker, b"1") {
-        log::warn!("Could not write autostart marker: {e}");
-        return false;
-    }
-    false
 }
 
 /// Tauri command — the pending-update banner's "Check for updates" CTA.
@@ -1944,7 +1650,7 @@ pub fn run() {
             cowork_commands::cowork_set_lan_ip_override,
             cowork_commands::cowork_retry_admin_elevation,
             restart_sidecar,
-            get_startup_rejection,
+            startup_rejection::get_startup_rejection,
             pending_update::get_pending_update_hint,
             check_for_update_now,
             show_in_file_manager,
@@ -4562,540 +4268,4 @@ mod classify_opened_url_tests {
         );
     }
 
-}
-
-/// Tests for the startup-file rejection surfacing (issue #630): the path-free
-/// reason-code mapping and the buffered-rejection take/clear semantics.
-#[cfg(test)]
-mod startup_rejection_tests {
-    use super::*;
-    use std::path::PathBuf;
-    use std::sync::Mutex as StdMutex;
-
-    // Serialize tests that mutate STARTUP_REJECTION (a process-wide static).
-    static REJECTION_LOCK: StdMutex<()> = StdMutex::new(());
-
-    #[test]
-    fn reason_code_is_stable_and_path_free() {
-        // The codes are the cross-process contract with the App.svelte toast
-        // map — assert the exact strings so a rename can't silently desync.
-        assert_eq!(
-            rejection_reason_code(&RejectionReason::UnsupportedExtension {
-                ext: "exe".into(),
-                path: PathBuf::from("/secret/place/file.exe"),
-            }),
-            "unsupported-extension"
-        );
-        assert_eq!(
-            rejection_reason_code(&RejectionReason::NotAFile {
-                path: PathBuf::from("/secret/place/dir"),
-            }),
-            "not-a-file"
-        );
-        assert_eq!(
-            rejection_reason_code(&RejectionReason::SuspiciousColon {
-                path: PathBuf::from("/secret/place/file.md:Zone.Identifier"),
-                index: 7,
-            }),
-            "suspicious-path"
-        );
-    }
-
-    #[test]
-    fn reason_code_never_leaks_the_path() {
-        // The reason code is a fixed enum string; assert it shares no substring
-        // with a path that would be sensitive to leak into a DOM toast.
-        let secret = "/Users/victim/Secret Plans.md";
-        let code = rejection_reason_code(&RejectionReason::NotAFile {
-            path: PathBuf::from(secret),
-        });
-        assert!(
-            !code.contains("Secret") && !code.contains('/'),
-            "reason code must not embed the rejected path"
-        );
-    }
-
-    #[test]
-    fn buffer_then_get_takes_once_then_returns_none() {
-        let _g = REJECTION_LOCK.lock().unwrap();
-        clear_startup_rejection();
-
-        buffer_startup_rejection_code(rejection_reason_code(&RejectionReason::UnsupportedExtension {
-            ext: "exe".into(),
-            path: PathBuf::from("/x/file.exe"),
-        }));
-        assert_eq!(
-            get_startup_rejection(),
-            Some("unsupported-extension".to_string()),
-            "first poll returns the buffered code"
-        );
-        assert_eq!(
-            get_startup_rejection(),
-            None,
-            "the buffer is TAKEN — a second poll (e.g. WebView reload) returns None"
-        );
-    }
-
-    #[test]
-    fn clear_drops_a_buffered_rejection() {
-        let _g = REJECTION_LOCK.lock().unwrap();
-        clear_startup_rejection();
-
-        buffer_startup_rejection_code(rejection_reason_code(&RejectionReason::NotAFile {
-            path: PathBuf::from("/x/missing.md"),
-        }));
-        // restart_sidecar calls clear_startup_rejection — a stale rejection from
-        // the previous launch must not survive into the next init-time drain.
-        clear_startup_rejection();
-        assert_eq!(
-            get_startup_rejection(),
-            None,
-            "clear must drop the buffered rejection"
-        );
-    }
-
-    #[test]
-    fn buffer_is_last_write_wins() {
-        let _g = REJECTION_LOCK.lock().unwrap();
-        clear_startup_rejection();
-
-        buffer_startup_rejection_code(rejection_reason_code(&RejectionReason::NotAFile {
-            path: PathBuf::from("/x/a.md"),
-        }));
-        buffer_startup_rejection_code(rejection_reason_code(&RejectionReason::UnsupportedExtension {
-            ext: "exe".into(),
-            path: PathBuf::from("/x/b.exe"),
-        }));
-        assert_eq!(
-            get_startup_rejection(),
-            Some("unsupported-extension".to_string()),
-            "the most recent buffered reason wins"
-        );
-        clear_startup_rejection();
-    }
-
-    /// Un-gated since #1344 (`opened_url_reason_code` is now unconditionally
-    /// compiled), so this contract runs on ubuntu, windows AND macOS.
-    #[test]
-    fn opened_url_reason_codes_are_path_free_and_stable() {
-        assert_eq!(
-            opened_url_reason_code(&OpenedUrlRejection::NonFileScheme),
-            "non-file-url"
-        );
-        assert_eq!(
-            opened_url_reason_code(&OpenedUrlRejection::NonEmptyHost),
-            "suspicious-path"
-        );
-        assert_eq!(
-            opened_url_reason_code(&OpenedUrlRejection::ConversionFailed),
-            "not-a-file"
-        );
-
-        // The delegating arm reuses the argv path's codes rather than minting
-        // new ones — App.svelte's reason-code→message map already handles both.
-        let unsupported = opened_url_reason_code(&OpenedUrlRejection::PathRejected(
-            RejectionReason::UnsupportedExtension {
-                ext: "exe".into(),
-                path: PathBuf::from("/secret/place/file.exe"),
-            },
-        ));
-        assert_eq!(unsupported, "unsupported-extension");
-        let not_a_file = opened_url_reason_code(&OpenedUrlRejection::PathRejected(
-            RejectionReason::NotAFile {
-                path: PathBuf::from("/Users/victim/Secret Plans.md"),
-            },
-        ));
-        assert_eq!(not_a_file, "not-a-file");
-
-        // Mirrors `reason_code_never_leaks_the_path`: delegating must never
-        // start putting a path on the wire.
-        for code in [unsupported, not_a_file] {
-            assert!(
-                !code.contains('/') && !code.contains("Secret"),
-                "delegated reason code must not embed the rejected path"
-            );
-        }
-    }
-
-    /// THE delivery-ordering contract, and the reason `surface_startup_rejection`
-    /// was split around a closure.
-    ///
-    /// Without this, two one-line mutations reverted the #1344 fix with a fully
-    /// green suite: deleting the buffer write made every path emit-only again
-    /// (the original cold-start drop), and swapping buffer/nudge let a live
-    /// listener drain an empty slot (the nudge is payload-free, so there is
-    /// nothing to fall back on). Both are killed here by asserting from INSIDE
-    /// the nudge that the code is already readable.
-    #[test]
-    fn the_code_is_buffered_before_the_nudge_fires() {
-        let _g = REJECTION_LOCK.lock().unwrap();
-        clear_startup_rejection();
-
-        let mut nudged = false;
-        surface_startup_rejection_with("unsupported-extension", || {
-            nudged = true;
-            // A real listener answers the nudge by calling get_startup_rejection.
-            // Doing that here is the whole point: it must already be there.
-            assert_eq!(
-                get_startup_rejection(),
-                Some("unsupported-extension".to_string()),
-                "the buffer must be populated BEFORE the nudge is emitted, or a \
-                 listener that answers immediately drains nothing"
-            );
-        });
-        assert!(nudged, "the nudge must actually be invoked");
-        clear_startup_rejection();
-    }
-
-    /// A failed emit must not cost the toast — the code stays queued for the
-    /// client's next init-time drain. This is the justification printed in
-    /// `surface_startup_rejection`'s `log::warn!`, now enforced.
-    #[test]
-    fn a_failed_nudge_leaves_the_code_buffered() {
-        let _g = REJECTION_LOCK.lock().unwrap();
-        clear_startup_rejection();
-
-        surface_startup_rejection_with("not-a-file", || { /* emit failed; do nothing */ });
-
-        assert_eq!(
-            get_startup_rejection(),
-            Some("not-a-file".to_string()),
-            "a dropped nudge costs one event, not the toast"
-        );
-        clear_startup_rejection();
-    }
-
-    /// One OS batch yields exactly one buffered code, so the toast the user gets
-    /// does not depend on where the client's async drain happened to land.
-    #[test]
-    fn a_batch_collapses_to_one_deterministic_code() {
-        let mut empty = RejectionBatch::default();
-        assert_eq!(empty.resolve(), None, "nothing rejected, nothing to surface");
-        empty.record("unsupported-extension");
-        assert_eq!(
-            empty.resolve(),
-            Some("unsupported-extension"),
-            "a lone rejection keeps its specific reason"
-        );
-
-        let mut mixed = RejectionBatch::default();
-        mixed.record("unsupported-extension");
-        mixed.record("not-a-file");
-        assert_eq!(
-            mixed.resolve(),
-            Some("multiple-rejected"),
-            "a mixed batch has no single true reason, so it must not claim one"
-        );
-
-        let mut same = RejectionBatch::default();
-        same.record("not-a-file");
-        same.record("not-a-file");
-        assert_eq!(
-            same.resolve(),
-            Some("multiple-rejected"),
-            "multiplicity is reported even when the reason agrees — four of five \
-             files opening is the case a singular message misdescribes"
-        );
-    }
-
-    // ---- #1416: the post-validation failure code and its batch ------------
-    //
-    // No statics are touched here, so these need neither `REJECTION_LOCK` nor
-    // `pending_opens_tests`'s `FLAG_LOCK`.
-
-    #[test]
-    fn open_failed_code_is_stable() {
-        // The cross-process contract with `messageForStartupRejection`'s
-        // explicit `case "open-failed"`. A rename on either side desyncs
-        // silently, because the client's `default` renders the same text.
-        assert_eq!(CODE_OPEN_FAILED, "open-failed");
-    }
-
-    #[test]
-    fn deferred_codes_are_stable_and_distinct_from_the_failure_codes() {
-        assert_eq!(CODE_OPEN_DEFERRED, "open-deferred");
-        assert_eq!(CODE_MULTIPLE_DEFERRED, "multiple-deferred");
-        // The split is the point: a retained queue still opens on the next
-        // successful start, so it must never render as the past-tense failure
-        // message. Collapsing these back onto the failure codes would restore
-        // the false statement this pair exists to remove.
-        assert_ne!(CODE_OPEN_DEFERRED, CODE_OPEN_FAILED);
-        assert_ne!(CODE_MULTIPLE_DEFERRED, CODE_MULTIPLE_REJECTED);
-    }
-
-    /// Drive `post_paths_and_surface` with an injected poster.
-    ///
-    /// `tauri::async_runtime::block_on` works in a plain `#[test]` — it lazily
-    /// initialises its own runtime and needs no `AppHandle`, no `#[tokio::test]`
-    /// and no tokio dev-dependency.
-    fn run_batch(paths: &[&str], failing: &[&str], seed: Option<&'static str>) -> Vec<&'static str> {
-        let mut surfaced: Vec<&'static str> = Vec::new();
-        let mut batch = RejectionBatch::default();
-        if let Some(code) = seed {
-            batch.record(code);
-        }
-        // Real screened paths, because `post_paths_and_surface` now takes
-        // `ScreenedOpenPath` and the newtype has no test-only constructor by
-        // design (#1415). The names are matched on the file name below, so the
-        // tempdir prefix is invisible to every caller.
-        let dir = tempfile::TempDir::new().expect("tempdir");
-        let owned: Vec<ScreenedOpenPath> = paths
-            .iter()
-            .map(|name| {
-                let path = dir.path().join(name);
-                std::fs::write(&path, b"x").expect("write fixture");
-                validate_open_candidate(path).expect("fixture must pass the screener")
-            })
-            .collect();
-        tauri::async_runtime::block_on(post_paths_and_surface(
-            "test",
-            owned,
-            batch,
-            // The `fail` decision is computed OUTSIDE the async block on
-            // purpose: an `Fn` closure may only borrow, so an `async move` that
-            // captured `failing` itself would be E0507 on the second call.
-            |path| {
-                let name = path
-                    .as_path()
-                    .file_name()
-                    .map(|n| n.to_string_lossy().into_owned())
-                    .unwrap_or_default();
-                let fail = failing.iter().any(|f| name == *f);
-                async move {
-                    if fail {
-                        Err("boom".to_string())
-                    } else {
-                        Ok(())
-                    }
-                }
-            },
-            |code| surfaced.push(code),
-        ));
-        surfaced
-    }
-
-    #[test]
-    fn a_fully_successful_batch_surfaces_nothing() {
-        assert!(
-            run_batch(&["a.md", "b.md"], &[], None).is_empty(),
-            "opening two files successfully must not toast"
-        );
-    }
-
-    #[test]
-    fn a_failed_post_surfaces_the_open_failed_code_once() {
-        // The #1416 bug itself: this used to be a `log::warn!` and nothing else,
-        // which in a release build is a file the user never opens.
-        assert_eq!(run_batch(&["big.md"], &["big.md"], None), vec!["open-failed"]);
-        // Two failures are still ONE surface call, and report multiplicity.
-        assert_eq!(
-            run_batch(&["a.md", "b.md", "c.md"], &["a.md", "c.md"], None),
-            vec!["multiple-rejected"]
-        );
-    }
-
-    #[test]
-    fn a_validation_rejection_and_a_post_failure_share_one_toast() {
-        // One Finder multi-select: a .pdf refused by the validator and a 60 MB
-        // .md refused by the server. Two surface calls would write twice into
-        // the one-slot buffer, and `useNotifications` would show a count badge
-        // whose value depends on where the client's async drain landed — the
-        // exact race `RejectionBatch` exists to remove.
-        let surfaced = run_batch(&["huge.md"], &["huge.md"], Some("unsupported-extension"));
-        assert_eq!(
-            surfaced,
-            vec!["multiple-rejected"],
-            "the batch spans validation AND delivery, and resolves exactly once"
-        );
-    }
-
-    #[test]
-    fn a_batch_with_nothing_to_post_still_surfaces_its_rejections() {
-        // A fully-rejected batch (the common shape since #1344: a double-clicked
-        // .pdf, a stale alias) has nothing to POST and still has something to
-        // say. `post_batch_for_app`'s early return mirrors this by surfacing
-        // BEFORE it returns — that ordering is not covered here, since it needs
-        // an AppHandle; this pins the helper's half of it.
-        assert_eq!(run_batch(&[], &[], Some("not-a-file")), vec!["not-a-file"]);
-    }
-}
-
-#[cfg(test)]
-mod autostart_tests {
-    use super::*;
-
-    #[test]
-    fn detects_the_flag_anywhere_after_argv0() {
-        assert!(is_autostart_launch(&[
-            "tandem".into(),
-            AUTOSTART_FLAG.into()
-        ]));
-        assert!(is_autostart_launch(&[
-            "tandem".into(),
-            "--other".into(),
-            AUTOSTART_FLAG.into(),
-        ]));
-    }
-
-    #[test]
-    fn absent_flag_is_a_normal_launch() {
-        assert!(!is_autostart_launch(&["tandem".into()]));
-        assert!(!is_autostart_launch(&["tandem".into(), "doc.md".into()]));
-        // Prefix/suffix collisions must not match — exact comparison only.
-        assert!(!is_autostart_launch(&[
-            "tandem".into(),
-            "--tandem-autostart-please".into(),
-        ]));
-        assert!(!is_autostart_launch(&["tandem".into(), "-tandem-autostart".into()]));
-    }
-
-    #[test]
-    fn argv0_is_never_read_as_the_flag() {
-        // An executable renamed to the flag string must not self-trigger.
-        assert!(!is_autostart_launch(&[AUTOSTART_FLAG.into()]));
-    }
-
-    #[test]
-    fn env_kill_switch_downgrades_to_a_normal_launch() {
-        let args = vec!["tandem".to_string(), AUTOSTART_FLAG.to_string()];
-        assert!(resolve_autostart_launch(&args, None));
-        assert!(resolve_autostart_launch(&args, Some("0")));
-        assert!(resolve_autostart_launch(&args, Some("")));
-        assert!(!resolve_autostart_launch(&args, Some("1")));
-        // The kill switch can only ever downgrade — it never invents an
-        // autostart launch out of a normal one.
-        let plain = vec!["tandem".to_string()];
-        assert!(!resolve_autostart_launch(&plain, Some("1")));
-        assert!(!resolve_autostart_launch(&plain, None));
-    }
-
-    #[test]
-    fn hides_only_when_autostart_and_a_tray_exists() {
-        assert!(should_start_hidden(true, true));
-        // The trapdoor case: hiding here would leave an unreachable process
-        // holding :3478/:3479 with no tray icon and no window to close.
-        assert!(!should_start_hidden(true, false));
-        assert!(!should_start_hidden(false, true));
-        assert!(!should_start_hidden(false, false));
-    }
-
-    #[test]
-    fn autostart_flag_is_not_a_file_arg() {
-        // extract_file_arg skips `-`-prefixed args, so the flag can never be
-        // resolved as a path. Pin it rather than relying on it by accident.
-        let cwd = std::path::Path::new("/tmp");
-        let args = vec!["tandem".to_string(), AUTOSTART_FLAG.to_string()];
-        assert!(matches!(extract_file_arg(&args, cwd), Ok(None)));
-    }
-
-    #[test]
-    fn autostart_flag_does_not_shadow_a_real_file_arg() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let file = dir.path().join("doc.md");
-        std::fs::write(&file, b"# hi").expect("write");
-
-        let args = vec![
-            "tandem".to_string(),
-            AUTOSTART_FLAG.to_string(),
-            file.to_string_lossy().to_string(),
-        ];
-        let resolved = extract_file_arg(&args, dir.path()).expect("should resolve");
-        assert_eq!(resolved.as_deref(), Some(file.as_path()));
-    }
-
-    #[test]
-    fn autostart_flag_alongside_a_bad_extension_still_rejects() {
-        // The flag must not mask the existing rejection path.
-        let dir = tempfile::tempdir().expect("tempdir");
-        let file = dir.path().join("payload.exe");
-        std::fs::write(&file, b"x").expect("write");
-
-        let args = vec![
-            "tandem".to_string(),
-            AUTOSTART_FLAG.to_string(),
-            file.to_string_lossy().to_string(),
-        ];
-        assert!(matches!(
-            extract_file_arg(&args, dir.path()),
-            Err(RejectionReason::UnsupportedExtension { .. })
-        ));
-    }
-
-    // Serialize the tests that mutate LAUNCHER_DEFERRED (a process-wide static).
-    static DEFERRAL_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-    #[test]
-    fn deferral_latch_releases_exactly_once() {
-        let _guard = DEFERRAL_LOCK.lock().unwrap_or_else(|p| p.into_inner());
-        LAUNCHER_DEFERRED.store(true, Ordering::Release);
-
-        // `swap` is what makes concurrent shows safe: only the first caller
-        // sees `true`, so the start POST can never be issued twice.
-        assert!(LAUNCHER_DEFERRED.swap(false, Ordering::AcqRel));
-        assert!(!LAUNCHER_DEFERRED.swap(false, Ordering::AcqRel));
-        assert!(!LAUNCHER_DEFERRED.load(Ordering::Acquire));
-    }
-
-    #[test]
-    fn deferral_latch_is_off_for_a_normal_launch() {
-        let _guard = DEFERRAL_LOCK.lock().unwrap_or_else(|p| p.into_inner());
-        // Mirrors setup(): the latch is stored from the resolved autostart
-        // state, so a normal launch never defers and never posts.
-        LAUNCHER_DEFERRED.store(resolve_autostart_launch(&["tandem".into()], None), Ordering::Release);
-        assert!(!LAUNCHER_DEFERRED.load(Ordering::Acquire));
-        assert!(!LAUNCHER_DEFERRED.swap(false, Ordering::AcqRel));
-    }
-
-    #[test]
-    fn a_failed_release_re_arms_the_latch_for_the_next_signal() {
-        let _guard = DEFERRAL_LOCK.lock().unwrap_or_else(|p| p.into_inner());
-        LAUNCHER_DEFERRED.store(true, Ordering::Release);
-
-        // `note_user_presence` claims the latch with `swap` before it knows
-        // whether the POST will succeed. If the sidecar never came up, or the
-        // request failed, it must put the claim back — otherwise a transient
-        // failure permanently strands the launcher, which is exactly the bug
-        // the health-gating was added to prevent.
-        let claimed = LAUNCHER_DEFERRED.swap(false, Ordering::AcqRel);
-        assert!(claimed, "the first signal claims the latch");
-        assert!(!LAUNCHER_DEFERRED.load(Ordering::Acquire));
-
-        // ...release fails...
-        LAUNCHER_DEFERRED.store(true, Ordering::Release);
-
-        // ...so a later presence signal can still claim it.
-        assert!(LAUNCHER_DEFERRED.swap(false, Ordering::AcqRel));
-    }
-
-    #[test]
-    fn deferral_latch_survives_a_sidecar_restart_until_released() {
-        let _guard = DEFERRAL_LOCK.lock().unwrap_or_else(|p| p.into_inner());
-        LAUNCHER_DEFERRED.store(true, Ordering::Release);
-
-        // start_sidecar re-reads the latch on every spawn attempt, so a crash
-        // loop before the user opens the window keeps deferring...
-        for _attempt in 0..3 {
-            assert!(LAUNCHER_DEFERRED.load(Ordering::Acquire));
-        }
-        // ...and once released, a later restart does NOT re-defer. This is the
-        // regression a statically captured env var would have shipped.
-        LAUNCHER_DEFERRED.swap(false, Ordering::AcqRel);
-        for _attempt in 0..3 {
-            assert!(!LAUNCHER_DEFERRED.load(Ordering::Acquire));
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_marker_reports_unseen_once_then_seen() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let root = dir.path().join("nested/appdata");
-
-        // First autostart launch: no marker yet -> show the window once.
-        assert!(!autostart_seen_and_mark(&root));
-        assert!(root.join(AUTOSTART_SEEN_MARKER).exists());
-        // Every launch after that trusts the tray.
-        assert!(autostart_seen_and_mark(&root));
-        assert!(autostart_seen_and_mark(&root));
-    }
 }
